Simplify dependencies
This project used to require the Conan package manager in order for cmake to generate a Makefile. This was largely unnecessary since this project is so simple.

I have removed spdlog and fmt from the required dependencies since they could be replaced with the c++ stdlib directly. The STB Image library is still necessary for decoding different image formats. The integration for STB Image was switched to git submodules to remove the need for Conan.